+++ SWADroid/src/es/ugr/swad/swadroid/modules/Groups.java
@@ -26,6 +26,8 @@ import es.ugr.swad.swadroid.Constants;
 import es.ugr.swad.swadroid.R;
 import es.ugr.swad.swadroid.model.Group;
 import es.ugr.swad.swadroid.model.Model;
+import es.ugr.swad.swadroid.webservices.SOAPClient;
+
 import org.ksoap2.serialization.SoapObject;
 import org.xmlpull.v1.XmlPullParserException;
 
@@ -109,7 +111,7 @@ public class Groups extends Module {
     @Override
     protected void requestService() throws NoSuchAlgorithmException,
             IOException, XmlPullParserException {
-        createRequest();
+    	createRequest(SOAPClient.CLIENT_TYPE);
         addParam("wsKey", Constants.getLoggedUser().getWsKey());
         addParam("courseCode", (int) courseCode);
         sendRequest(Group.class, false);